login/register

+++ project-four/project-django-app/comments/views.py
@@ -1,13 +1,19 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
 from rest_framework import status
+from rest_framework.exceptions import NotFound, PermissionDenied
+
 
 from .serializers.common import CommentSerializer
 from .models import Comment
 
 
 class CommentListView(APIView):
+    permission_classes = (IsAuthenticated,)
+
     def post(self, request):
+        request.data["owner"] = request.user.id
         comment_to_create = CommentSerializer(data=request.data)
         if comment_to_create.is_valid():
             comment_to_create.save()
@@ -19,9 +25,9 @@ class CommentDetailView(APIView):
     def delete(self, _request, pk):
         try:
             comment_to_delete = Comment.objects.get(pk=pk)
-        except comment_to_delete.DoesNotExist:
-            return Response(
-                comment_to_delete.errors, status=status.HTTP_422_UNPROCESSABLE_ENTITY
-            )
+        except Comment.DoesNotExist:
+            raise NotFound()
+        if comment_to_delete.owner != request.user:
+            raise PermissionDenied()
         comment_to_delete.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
